Fixed some dependancies

+++ DystopiaEngine/DystopiaEngine/Include/Component/RigidBody.h
@@ -74,7 +74,7 @@ namespace Dystopia
 	public:
 		// ===================================== MEMBER FUNCTIONS ==================================== // 
 		void Integrate(float _dt);
-		void UpdateResult(double alpha) const;
+		void UpdateResult(float alpha) const;
 
 		// void Update(float _dt);
 		// void LateUpdate(float _dt);
